+++ python/mujoco/bindings_test.py
@@ -994,6 +994,18 @@ Euler integrator, semi-implicit in velocity.
         bodyexclude=0,
         geomid=geomid)
 
+  def test_mju_box_qp(self):
+    n = 5
+    res = np.zeros(n)
+    r = np.zeros((n, n+7))
+    index = np.zeros(n, np.int32)
+    h = np.eye(n)
+    g = np.ones((n,))
+    lower = -np.ones((n,))
+    upper = np.ones((n,))
+    rank = mujoco.mju_boxQP(res, r, index, h, g, lower, upper)
+    self.assertGreater(rank, -1)
+
   @parameterized.product(flg_html=(False, True), flg_pad=(False, True))
   def test_mj_printSchema(self, flg_html, flg_pad):  # pylint: disable=invalid-name
     # Make sure that mj_printSchema doesn't raise an exception